Create Street Fighter 2 Mix (Ver 0.98c 201214)
This is a hack of Street Fighter II': Champion Edition. For more information about this hack go to https://sf2mix.github.io/. The hack is perfectly playable on jotego's core except for Complete Mode. The fake 3d backgrounds in Complete Mode are (currently) not supported by jotego's core. Therefore the default dip switch setting for Complete Mode is off.