Update data.json
Added extension "ngc" so that backwards-compatible NGPC ROMs will be viewable from within the NGP Pocket core file browser. Far more of these colour games are compatible than on the MAME published list on the core readme. 

Some may not load correctly but this at least gives the user the freedom to try. If this needs to be explained in docs then that will be done.